- Extract cluster readiness test - Add check in BeforeAll
